Propagate github credentials using volumes instead of docker-compose environment variables
Issue: When recreating the AP, host files in other containers were not
updated and the connection with the AP was lost.
Restarting everything with `docker-compose up -d` didn't work because
the command was started from the jupyter container itself and was
killed during the process of restart.

Simply restarting the AP doesn't refresh environment variables
specified in the docker-compose file.

Solution: overwrite the configuration file using a volume shared
between the jupyter container and the AP. Therefore, simply restarting
the container will refresh the environment and the AP credentials.